Move the apex home onto the single [locale] tree
Serve the landing page from [locale]/page.tsx and route the apex through the same
locale middleware: the naked home (/) rewrites to the default-locale branch (/en)
while the URL stays "/", an explicit /en redirects back to /, and a localized home
(/hu) passes through. Locale homes join the cacheable set alongside the naked apex.
