+++ app/tests/integration/home.test.tsx
@@ -1,4 +1,4 @@
-import RootPage from "@/app/(hybrid)/page";
+import RootPage from "@/app/(hybrid)/[locale]/page";
 import { render, screen } from "@testing-library/react";
 
 // Mock the server-side auth check
